# crates/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OVLive API",
    "version": "0.1.0",
...
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key: ovl_..."
      },
      "Basic": {
        "type": "http",
        "scheme": "basic",
        "description": "Account email + password"
      }
    }
  },
  "paths": {
...
    "/v1/vehicles/{id}": {
      "get": {
        "summary": "Vehicle detail (route shape, upcoming stops, predicted next trip)",
        "description": "Includes `next_trip` (`{ line_public_number, destination, start_unix }` or null) \u2014 the line this vehicle becomes next, from KV78Turbo block/omloop chaining. The WS `VehicleState` carries the same as `next_line_public_number` / `next_destination` / `next_start_unix`.",
        "security": [
          {},
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "\"<dataowner>:<vehicle_number>\"",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/v1/stops": {
      "get": {
        "summary": "Transit stops in a viewport",
        "description": "**Deprecated \u2014 will be removed.** Byte-compatible shim for consumers of the pre-Rust OVLive API: camelCase keys, capitalised enums (`\"Bus\"`, `\"OnStop\"`), `yyyyMMdd` `operatingDay`, and plain-text (not JSON) error bodies. Responses carry `Deprecation: true`. Returns `{stops:[{stopId,stopName,lat,lon}]}`, capped by `MAX_STOPS_RESULTS` (default 500). Stations and gtfs-nl `stoparea:*` rows are excluded, as before. Ordering is nearest-to-viewport-centre; the old SQL had no `ORDER BY`. No replacement yet.",
        "deprecated": true,
        "tags": [
          "Deprecated"
        ],
        "security": [
          {},
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "neLat",
            "in": "query",
            "description": "Northeast latitude (required)",
            "schema": {
              "type": "number"
            },
            "example": 52.2
          },
          {
            "name": "neLon",
            "in": "query",
            "description": "Northeast longitude (required)",
            "schema": {
              "type": "number"
            },
            "example": 5.1
          },
          {
            "name": "swLat",
            "in": "query",
            "description": "Southwest latitude (required)",
            "schema": {
              "type": "number"
            },
            "example": 51.8
          },
          {
            "name": "swLon",
            "in": "query",
            "description": "Southwest longitude (required)",
            "schema": {
              "type": "number"
            },
            "example": 4.3
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "`invalid neLat` / `invalid viewport bounds` / `viewport too large` (plain text). Area is capped by `MAX_VIEWPORT_AREA`, default 2.0 deg\u00b2."
          },
          "500": {
            "description": "`internal error` (plain text) \u2014 feed not loaded yet"
          },
          "401": {
            "description": "Invalid/revoked key (only when a key is supplied)"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
  },
...
}